requirements.txt
zep-cloud>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
rich>=13.0.0
requests>=2.31.0
uuid
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
neo4j>=5.0.0
jellyfish>=1.2.0
flask>=2.3.0
flask-cors>=4.0.0
